remove `SubqueryFilterToJoin`
